refactor(ivr_node): icon component usage
