+++ recipes/cards/ivr_node/ivr_node.mdx
@@ -111,7 +111,7 @@ A node in IVR used to integration DX expert view in iFrame to create complex flo
   <dt-button importance="clear" icon-position="right">
     Launch Expert
     <template #icon>
-      <icon-launch />
+      <dt-icon name="external-link" />
     </template>
   </dt-button>
 </dt-recipe-ivr-node>
