+++ recipes/cards/ivr_node/ivr_node_constants.js
@@ -1,12 +1,3 @@
-import IconDialpad from '@dialpad/dialtone/lib/dist/vue/icons/IconDialpad';
-import IconPhoneDialer from '@dialpad/dialtone/lib/dist/vue/icons/IconPhoneDialer';
-import IconVolumeUp from '@dialpad/dialtone/lib/dist/vue/icons/IconVolumeUp';
-import IconPhoneTransfer from '@dialpad/dialtone/lib/dist/vue/icons/IconPhoneTransfer';
-import IconCallEndArrow from '@dialpad/dialtone/lib/dist/vue/icons/IconCallEndArrow';
-import IconCallMerge from '@dialpad/dialtone/lib/dist/vue/icons/IconCallMerge';
-import IconExpertNode from '@dialpad/dialtone/lib/dist/vue/icons/IconExpertNode';
-import IconBranch from '@dialpad/dialtone/lib/dist/vue/icons/IconBranch';
-
 export const IVR_NODE_PROMPT_MENU = 'promptmenu';
 export const IVR_NODE_PROMPT_COLLECT = 'promptcollect';
 export const IVR_NODE_PROMPT_PLAY = 'promptplay';
@@ -17,14 +8,14 @@ export const IVR_NODE_TRANSFER = 'transfer';
 export const IVR_NODE_HANGUP = 'hangup';
 
 export const IVR_NODE_ICON_TYPES = {
-  [IVR_NODE_PROMPT_MENU]: IconDialpad,
-  [IVR_NODE_PROMPT_COLLECT]: IconPhoneDialer,
-  [IVR_NODE_PROMPT_PLAY]: IconVolumeUp,
-  [IVR_NODE_EXPERT]: IconExpertNode,
-  [IVR_NODE_BRANCH]: IconBranch,
-  [IVR_NODE_GO_TO]: IconCallMerge,
-  [IVR_NODE_TRANSFER]: IconPhoneTransfer,
-  [IVR_NODE_HANGUP]: IconCallEndArrow,
+  [IVR_NODE_PROMPT_MENU]: 'keypad',
+  [IVR_NODE_PROMPT_COLLECT]: 'dialer',
+  [IVR_NODE_PROMPT_PLAY]: 'volume-2',
+  [IVR_NODE_EXPERT]: 'corner-down-right',
+  [IVR_NODE_BRANCH]: 'network',
+  [IVR_NODE_GO_TO]: 'chevrons-up',
+  [IVR_NODE_TRANSFER]: 'transfer',
+  [IVR_NODE_HANGUP]: 'phone-hang-up',
 };
 
 export const IVR_NODE_LABELS = {
